langchain_project/pyproject.toml
[project]
name = "langchain-project"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "langchain>=1.0.0",
    "langchain-deepseek>=0.1.4",
    "langgraph>=0.6.7",
    "load-dotenv>=0.1.0",
]

[[tool.uv.index]]
url = "https://pypi.tuna.tsinghua.edu.cn/simple"
default = true
